+++ tests/testthat.R
@@ -1,4 +1,8 @@
 library(testthat)
 library(gh)
 
-test_check("gh")
+# Don't want to use keyrings on CRAN
+withr::with_envvar(
+  c(GH_NO_KEYRING = "true"),
+  test_check("gh")
+)