+++ NEWS.md
@@ -13,6 +13,10 @@
    `Accept` HTTP header (#91).
 -  New `gh_gql()` function to make it easier to work with the GitHub
    GraphQL API.
+-  gh now supports separate personal access tokens for GitHub Enterprise
+   sites.
+-  gh now supports storing your GitHub personal access tokens (PAT) in the
+   system keyring, via the keyring package.
 
 # 1.0.1